docs/projeto/api/requirements.txt
# lista de bibliotecas Python

fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
requests
python-dotenv
